test: fold vertex service_tier regression test into mapped cost calculator tests
